tests: Move selected snapshot coverage into data
Selected-file snapshot persistence now lives in the data package, while
diff parser tests cover parser behavior only.

The snapshot tests still live in the core parser test module. That layout
makes the test suite describe snapshot persistence as parser behavior after
the code has moved to selected-change state.

This commit continues the selected-snapshot boundary split by moving the
snapshot persistence tests into the data test package.

The final commit will add a guard that keeps runtime snapshot IO out of the
parser module.
